deepen: fold Onboarding Fallback Notice into the Onboarding First Email module
The onboarding-first-email contract had two halves living in two
files: `lib/onboarding/cached-summary-delivery.ts` (the cached-summary
happy path) and `lib/email/onboarding-fallback-service.ts` (the long-
tail 'we're watching your tickers' notice for users with no cached
summaries across their tickers). Both paths checked + wrote the same
`User.onboardingFirstEmailSentAt` idempotency column. The only
production caller (`completeOnboardingBatched` in
`app/(auth)/onboarding/actions.ts`) had to dispatch between them
manually by inspecting the cached path's `DeliveryResult.reason` —
testing past the seam, not behaviour.

Deletion test: removing `onboarding-fallback-service.ts` and
absorbing the fallback as a private helper concentrates the
onboarding-first-email logic in one module behind a single
`sendOnboardingFirstEmail({userId, email, recipientName,
trackedTickers})` function. The caller stops branching; the seam
moves from inside the caller to inside the module.

Dependency category: in-process (Prisma + email transport are both
already injected through their own modules). No port/adapter
required — the fallback helper becomes private to the deepened
module's implementation.

Interface change for the caller:
  - was: import `deliverFirstOnboardingEmail` + `sendOnboardingFallbackNotice`,
         call the first, branch on `reason ∈ {'no_cached_summaries',
         'no_tickers'}`, then call the second.
  - now: import `sendOnboardingFirstEmail`, call once. The module
         decides which path applies.

Tests deleted: `__tests__/lib/email/onboarding-fallback-service.test.ts`
(211 LOC, 12 tests asserting on internal coercion of the now-private
helper).

Tests added: 13 new orchestrator tests in
`__tests__/lib/onboarding/cached-summary-delivery.test.ts` exercising
both branches through the `sendOnboardingFirstEmail` interface,
including: idempotency from the cached path skips fallback;
fallback fires only on `no_cached_summaries`/`no_tickers` (not on
`email_failed` / `internal_error`); fallback subject truncation; the
`onboarding-fallback-notice` tag; the `onboardingFirstSummaryId`
contract (only set on cached path). Suite: 27 → 40 tests, all
passing.

`pickBestSummaryForUser` and `deliverFirstOnboardingEmail` remain
exported so the legacy `/api/onboarding?action=deliver-summaries`
route (via the `deliverCachedSummaries` shim) and the existing test
surface keep working — the deepening adds the orchestrator on top
without breaking the cached-path interface.

CONTEXT.md gains the 'Onboarding First Email' domain term.
