+++ CONTEXT.md
@@ -24,6 +24,9 @@ A user's saved interest in a [Filing] type or ticker. Drives which [Summary] pay
 ### Onboarding
 The first-run flow that turns a new signup into an engaged subscriber: ticker selection, preview email, account creation, first cached [Summary] delivery.
 
+### Onboarding First Email
+The single email a freshly-onboarded user receives moments after completing [Onboarding]. Two paths, one contract: pick the best cached cross-user [Summary] across the user's tickers and send it; if no usable cached summary exists (long-tail unique-ticker case, or zero tickers), send the "we're watching your tickers" fallback notice instead. Implemented as one deep module (`lib/onboarding/cached-summary-delivery.ts`) behind a single `sendOnboardingFirstEmail({userId, email, recipientName, trackedTickers})` function — the caller never branches on which path actually fired. Idempotent on `User.onboardingFirstEmailSentAt`, the write-once marker shared by both paths so re-onboarding never double-sends regardless of which branch sent the original email. Replaces a former `lib/email/onboarding-fallback-service.ts` (`sendOnboardingFallbackNotice`) that the only caller (`completeOnboardingBatched` in `app/(auth)/onboarding/actions.ts`) had to dispatch to manually after inspecting the cached path's `reason` field; the split was a shallow seam (one adapter, one caller, both paths writing the same idempotency column) whose tests asserted on the seam's internal `DeliveryResult.reason` rather than on observable end-to-end "did the user get an email" behaviour. The deepened orchestrator is verified through the same `__tests__/lib/onboarding/cached-summary-delivery.test.ts` interface that already covers the cached path. The cached-summary mechanics (two-stage select, ranking, tiebreaker) remain the canonical implementation of `pickBestSummaryForUser` and `deliverFirstOnboardingEmail`, which the orchestrator delegates to internally.
+
 ### X Sentiment
 An enrichment of a [Summary] that captures public discussion on X (Twitter) about the ticker over a 7-day window. Direction (bullish/bearish/mixed/neutral/no_signal), shift, confidence, and citation-backed claims. Gated by form importance + a large-cap allowlist (S&P 500) to keep cost and pump-and-dump risk bounded. Runs only for paid-tier callers via the enrichment gate.
